requirements.txt
streamlit
transformers
sentencepiece
torch
argostranslate
requests
speechrecognition
gtts
pydub
numpy
deep-translator
docx
python-docx
PyPDF2
pyaudio
streamlit
bcrypt
langdetect
pdfplumber
google-generativeai